Use direct JOOQ codegen invocation
The codegen used to be invoked in compile phase using a specific profile
activation. As it turns out this is an awful idea. For example, it is
not possible to re-run the generation if something goes wrong because
your code won't compile anymore.

Fortunately the codegen can be invoked directly like this:

    mvn jooq-codegen:generate

The plugin config is read from pom.xml and everything works just as it
did before. Just a lot faster and more reliably.